Extract kubernetes mockups to a common py file
Some common mockups are used in the tests of all providers.
This PR prevents the duplication of code
